+++ CHANGELOG.rst
@@ -30,6 +30,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - Upgraded ska-mid-dish-ds-manager to v8.2.1
   - Implemented the ability to retake authority from DSC implicitly through commands that require authority.
   - Added a new command, RetakeAuthority, which can be used to retake authority from the DSC explicitly.
+  
+- Limited build state fetch to only positive transitions to communication status ESTABLISHED.
+- Allowed DS and B5DC build state information to be driven by change events.
 
 Version 9.4.0
 *************